Adding Iterative website
basic content is there although the full museum list will be added to the main page later from the json file
some minor css is added for clarity, but i would like to change to positions of the different sections of the page
functions that work in the search box include: borough checkboxes and search button, saturday button kind of works
the other function that currently works is the move to buttons at the bottom. All three buttons will work, but currently they're only set to work with the first museum in the list.